Refactor vite config
Not responsible for providing generate with the pages to prerender, split out virtual file plugin
