+++ package.json
@@ -116,7 +116,6 @@
     "jest-runner-prettier": "^0.3.6",
     "jest-runner-remark": "^1.0.28",
     "jsdom": "^19.0.0",
-    "make-dir": "^3.1.0",
     "prettier": "^2.3.2",
     "rehype-minify-json-script": "^3.0.0",
     "rehype-minify-whitespace": "^5.0.0",
